+++ packages/kbn-build/package.json
@@ -1,10 +1,5 @@
 {
   "name": "@kbn/build",
-  "kibana": {
-    "build": {
-      "skip": true
-    }
-  },
   "main": "./dist/index.js",
   "version": "1.0.0",
   "license": "Apache-2.0",
