+++ packages/kbn-build/src/index.js
@@ -1,3 +1,6 @@
 export { run } from './cli';
-export { buildProductionProjects } from './production';
+export {
+  buildProductionProjects,
+  prepareExternalProjectDependencies,
+} from './production';
 export { transformDependencies } from './utils/package_json';
