+++ packages/kbn-build/src/production/__fixtures__/external_packages/with_kibana_link_deps/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "quux",
+  "version": "1.0.0",
+  "dependencies": {
+    "@kbn/foo": "link:../../kibana/packages/foo"
+  }
+}
