+++ packages/kbn-build/src/production/__fixtures__/external_packages/with_other_link_deps/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "quux",
+  "version": "1.0.0",
+  "dependencies": {
+    "bar": "link:../foo/packages/bar"
+  }
+}
