+++ packages/kbn-build/src/utils/package_json.js
@@ -15,6 +15,8 @@ export const createProductionPackageJson = pkgJson => ({
   dependencies: transformDependencies(pkgJson.dependencies),
 });
 
+export const isLinkDependency = depVersion => depVersion.startsWith('link:');
+
 /**
  * Replaces `link:` dependencies with `file:` dependencies. When installing
  * dependencies, these `file:` dependencies will be copied into `node_modules`
@@ -28,7 +30,7 @@ export function transformDependencies(dependencies = {}) {
   const newDeps = {};
   for (const name of Object.keys(dependencies)) {
     const depVersion = dependencies[name];
-    if (depVersion.startsWith('link:')) {
+    if (isLinkDependency(depVersion)) {
       newDeps[name] = depVersion.replace('link:', 'file:');
     } else {
       newDeps[name] = depVersion;
